The logic of storing data in the application memory has been moved to a separate interface - FilmorateRepository. Controllers use a specific implementation of the interface (@Autowired)
