+++ src/main/java/ru/yandex/practicum/filmorate/serviceRepo/FilmorateRepository.java
@@ -1,0 +1,15 @@
+package ru.yandex.practicum.filmorate.serviceRepo;
+
+import org.springframework.stereotype.Component;
+
+import java.util.Collection;
+
+@Component
+public interface FilmorateRepository<T> {
+
+    Collection<T> getAll();
+
+    T create(T entity);
+
+    T update(T entity);
+}
